% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admissibility_test.R
\name{admissibility_test}
\alias{admissibility_test}
\title{Automatically check for data entry issues}
\usage{
admissibility_test(plan_of_study)
}
\arguments{
\item{plan_of_study}{igraph object - An igraph object created using the create_plan_of_study function}
}
\value{
Printed warnings for potential data entry issues
}
\description{
This function takes in a plan of study and a course, then checks for potential data entry issues.
It will detect issues in formatting with the csv (such as notes creating empty rows), if there
are cycles in the network, and if pre- and corequisites are appropriately defined.
}
