% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_sequencing.R
\name{average_sequencing}
\alias{average_sequencing}
\title{Calculates the average sequencing in a program}
\usage{
average_sequencing(plan_of_study, expected_time_to_degree = NULL)
}
\arguments{
\item{plan_of_study}{igraph object - An igraph object created using the create_plan_of_study function}

\item{expected_time_to_degree}{Numeric - The term where students are expected to finish (often 8)}
}
\value{
Numeric - the average sequencing in the program
}
\description{
This function calculates the average sequencing in the program using the
delay factors of the courses. The second argument, expected_time_to_degree
is optional. If it is not NULL, the average sequencing will be for courses
extending the student's time to degree.
}
