% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocking_factor.R
\name{blocking_factor}
\alias{blocking_factor}
\title{Calculates the blocking factor of a course}
\usage{
blocking_factor(plan_of_study, course, include_coreqs = TRUE)
}
\arguments{
\item{plan_of_study}{igraph object - An igraph object created using the create_plan_of_study function}

\item{course}{Numeric (vertex id) or String - The course to calculate the blocking factor of}

\item{include_coreqs}{logical - Indicates whether corequisites should be included in the calculation}
}
\value{
Numeric - the blocking factor
}
\description{
This function takes in a plan of study and a course, then finds that course's blocking factor.
The value is the number of courses 'blocked' by failing the given course.
}
