% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_collapse.R
\name{core_collapse}
\alias{core_collapse}
\title{Calculates the core collapse sequence for a plan of study}
\usage{
core_collapse(plan_of_study)
}
\arguments{
\item{plan_of_study}{igraph object - An igraph object created using the create_plan_of_study function}
}
\value{
List of two items: (1) sequence - the core collapse sequence, (2) the associated network for each entry
}
\description{
This function takes in a plan of study network and constructs the "core collapse sequence."
The core collapse sequence progressively removes courses from the plan of study with increasing
prereq counts and calculates the proportion of courses deleted at each step. The process stops
when all of the vertices have been removed. A sequence that decreases quickly to zero typically
indicates that the network is generally uniform with its prereqs. A sequence with more erratic
values that does not settle to zero smoothly would imply more dense sets of prereqs.
}
