% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cruciality.R
\name{cruciality}
\alias{cruciality}
\title{Calculates the cruciality of a course}
\usage{
cruciality(plan_of_study, course, include_coreqs = TRUE)
}
\arguments{
\item{plan_of_study}{igraph object - An igraph object created using the create_plan_of_study function}

\item{course}{Numeric (vertex id) or String - The course to calculate the cruciality of}

\item{include_coreqs}{logical - Indicates whether corequisites should be included in the calculation}
}
\value{
Numeric - the course's cruciality
}
\description{
This function takes in a plan of study and a course, then finds that course's cruciality.
The value is the sum of the blocking and delay factors of the course
}
