% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curriculum_rigidity.R
\name{curriculum_rigidity}
\alias{curriculum_rigidity}
\title{Calculates the curriculum rigidity}
\usage{
curriculum_rigidity(plan_of_study)
}
\arguments{
\item{plan_of_study}{igraph object - An igraph object created using the create_plan_of_study function}
}
\value{
Numeric - the curriculum rigidity
}
\description{
This function takes in a plan of study and then finds the curriculum's rigidity.
The rigidity is the beta index of the graph, which is the number of prerequisites
divided by the number of courses
}
