% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deferment_factor.R
\name{deferment_factor}
\alias{deferment_factor}
\title{Calculates the deferment factor of a course}
\usage{
deferment_factor(plan_of_study, course, expected_time_to_degree)
}
\arguments{
\item{plan_of_study}{igraph object - An igraph object created using the create_plan_of_study function}

\item{course}{Numeric (vertex id) or String - The course to calculate the deferment factor of}

\item{expected_time_to_degree}{Numeric - The term where students are expected to finish (often 8)}
}
\value{
Numeric - the deferment factor
}
\description{
This function takes in a plan of study and a course, then finds that course's deferment factor.
The value captures the number of terms the student can fail the course before extending their time
to degree.
}
