% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_factor.R
\name{delay_factor}
\alias{delay_factor}
\title{Calculates the delay factor of a course}
\usage{
delay_factor(plan_of_study, course, include_coreqs = TRUE)
}
\arguments{
\item{plan_of_study}{igraph object - An igraph object created using the create_plan_of_study function}

\item{course}{Numeric (vertex id) or String - The course to calculate the delay factor of}

\item{include_coreqs}{Logical - Calculates the delay factor using corequisites, default value is TRUE}
}
\value{
Numeric - the delay factor
}
\description{
This function takes in a plan of study and a course, then finds that course's delay factor.
The output is the longest path of prerequisites through the given course.
}
