% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explained_complexity.R
\name{explained_complexity}
\alias{explained_complexity}
\title{Calculates the explained complexity of courses extending time to degree}
\usage{
explained_complexity(
  plan_of_study,
  expected_time_to_degree,
  term_weighted = FALSE
)
}
\arguments{
\item{plan_of_study}{igraph object - An igraph object created using the create_plan_of_study function}

\item{expected_time_to_degree}{Numeric - The term where students are expected to finish (often 8)}

\item{term_weighted}{logical - TRUE if crucialities should be term-weighted}
}
\value{
Numeric - the explained complexity
}
\description{
This function takes in the subcomplexity graph from the transfer excess courses function, then finds the transfer delay
factor. The output is the proportion of complexity explained by courses extending time to degree.
}
