% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_inbound_courses.R
\name{find_inbound_courses}
\alias{find_inbound_courses}
\title{Find all possible prerequisites to a course}
\usage{
find_inbound_courses(plan_of_study, course)
}
\arguments{
\item{plan_of_study}{An igraph object created using the create_plan_of_study function}

\item{course}{The course to find all relevant prerequisites of}
}
\value{
An atomic vector of vertex ids for the course's prerequisites
}
\description{
This function takes in a plan of study and a course, then finds all the courses it is related to
through its prerequisties
}
