% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_number_of_free_terms.R
\name{find_number_of_free_terms}
\alias{find_number_of_free_terms}
\title{Finds the number of terms a course can move to}
\usage{
find_number_of_free_terms(
  plan_of_study,
  course,
  expected_time_to_degree,
  outbound_only
)
}
\arguments{
\item{plan_of_study}{igraph object - An igraph object created using the create_plan_of_study function}

\item{course}{Numeric (vertex id) or String - The course to find the number of free terms for}

\item{expected_time_to_degree}{Numeric - The term where students are expected to finish (often 8)}

\item{outbound_only}{logical - TRUE if the calculation should only include courses after the course (postreqs)}
}
\value{
Numeric - the number of free terms
}
\description{
This function takes in a plan of study, a course, the expected time to degree, and a
logical value indicating whether to restrict the number of free terms to those later
in the curriculum. The value is the number of free terms that the course can move to
}
