% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outbound_courses.R
\name{find_outbound_courses}
\alias{find_outbound_courses}
\title{Find all possible courses that depend on a particular course}
\usage{
find_outbound_courses(plan_of_study, course)
}
\arguments{
\item{plan_of_study}{An igraph object created using the create_plan_of_study function}

\item{course}{The course to find all relevant courses that directly or indirectly have it as a prereq}
}
\value{
An atomic vector of vertex ids for the course's following courses
}
\description{
This function takes in a plan of study and a course, then finds all the courses it is related to
through its prerequisites (after the course).
}
