% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflexibility_factor.R
\name{inflexibility_factor}
\alias{inflexibility_factor}
\title{Calculates inflexibility factor of a plan of study}
\usage{
inflexibility_factor(plan_of_study, time_to_degree)
}
\arguments{
\item{plan_of_study}{igraph object - An igraph object created using the create_plan_of_study function}

\item{time_to_degree}{numeric - expected time to degree, often 8}
}
\value{
list of (1) a dataframe of inflexibility factors and (2) a total inflexibility factor
}
\description{
Calculates the inflexibility factor for courses that have specific offering times extending chains beyond the
expected time to degree.
}
