% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_plan_of_study.R
\name{plot_plan_of_study}
\alias{plot_plan_of_study}
\title{Plots the plan of study with courses ordered by term}
\usage{
plot_plan_of_study(plan_of_study)
}
\arguments{
\item{plan_of_study}{igraph object - An igraph object created using the create_plan_of_study function}
}
\value{
Plots the plan of study in the 'plot' window
}
\description{
This function takes in a plan of study and plots it in the 'plot' window.
The courses are ordered horizontally by term and vertically by the outdegree
(i.e., number of prereqs) of the vertices in that column/term. The shading of
the nodes corresponds to the cruciality of the course. A darker blue indicates
higher cruciality while white indicates lower cruciality.
}
\details{
Note that there can be some overlap where a course is covering a path for a prereq,
which may make is seem like a course is a prereq for some other course when it is
in fact the course in a previous semester.
}
