% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reachability_factor.R
\name{reachability_factor}
\alias{reachability_factor}
\title{Calculates the reachability factor of a course}
\usage{
reachability_factor(plan_of_study, course, include_coreqs = TRUE)
}
\arguments{
\item{plan_of_study}{igraph object - An igraph object created using the create_plan_of_study function}

\item{course}{Numeric (vertex id) or String - The course to calculate the blocking factor of}

\item{include_coreqs}{logical - Indicates whether corequisites should be included in the calculation}
}
\value{
Numeric - the reachability factor
}
\description{
This function takes in a plan of study and a course, then finds that course's reachability factor.
The value is the number of courses needed to be passed before enrolling in the given course.
}
