% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_requisites.R
\name{simplify_requisites}
\alias{simplify_requisites}
\title{Convert requisites to original notation}
\usage{
simplify_requisites(requisites)
}
\arguments{
\item{requisites}{vector object - A vector describing the pre and corequisites (as strings)}
}
\value{
vector object -  A simplified vector describing the pre and corequisites (as strings)
}
\description{
This function takes in either the pre or corequisites of a plan of study as a vector, then
removes any additional information like OR relationships and minimum grades such that the
network can be analyzed using the traditional functions.
}
