% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_complexity.R
\name{structural_complexity}
\alias{structural_complexity}
\title{Calculates structural complexity of a plan of study}
\usage{
structural_complexity(plan_of_study)
}
\arguments{
\item{plan_of_study}{igraph object - An igraph object created using the create_plan_of_study function}
}
\value{
list of (1) a dataframe of course crucialities, delay factors, and blocking factors; (2) a numeric value
of structural complexity
}
\description{
This function takes in a plan of study, then finds the plan of study's structural complexity.
}
