% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcomplexity_graph.R
\name{subcomplexity_graph}
\alias{subcomplexity_graph}
\title{Creates a subcomplexity graph for a course}
\usage{
subcomplexity_graph(plan_of_study, course)
}
\arguments{
\item{plan_of_study}{igraph object - An igraph object created using the create_plan_of_study function}

\item{course}{Numeric (vertex id) or String - The course to find the subcomplexity graph of}
}
\value{
igraph object representing the course's subcomplexity graph.
}
\description{
This function takes in a plan of study and course, then constructs the subcomplexity graph for the course.
}
