% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_delay_factor.R
\name{transfer_delay_factor}
\alias{transfer_delay_factor}
\title{Calculates the transfer delay factor of a course}
\usage{
transfer_delay_factor(plan_of_study, expected_time_to_degree)
}
\arguments{
\item{plan_of_study}{igraph object - An igraph object created using the create_plan_of_study function}

\item{expected_time_to_degree}{Numeric - The term where students are expected to finish (often 8)}
}
\value{
Numeric - the transfer delay factor
}
\description{
This function takes in the subcomplexity graph from the transfer excess courses function, then finds the transfer delay
factor. The output is the sum of the longest paths of prerequisties through courses related to those beyond the
expected time to degree.
}
