% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_excess_courses.R
\name{transfer_excess_courses}
\alias{transfer_excess_courses}
\title{Finds the subcomplexity graph of courses beyond expected time to degree}
\usage{
transfer_excess_courses(
  plan_of_study,
  expected_time_to_degree,
  include_coreqs = TRUE
)
}
\arguments{
\item{plan_of_study}{igraph object - An igraph object created using the create_plan_of_study function}

\item{expected_time_to_degree}{Numeric - The term where students are expected to finish (often 8)}

\item{include_coreqs}{Logical - Calculates the delay factor using corequisites, default value is TRUE}
}
\value{
igraph object - the subcomplexity graph
}
\description{
This function takes in a plan of study and the expected time to degree, then outputs a subcomplexity
graph that contains all of the courses beyond the time to degree and their prerequisites.
}
