% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what_if.R
\name{what_if}
\alias{what_if}
\title{Calculate hypotheticals of adding/deleting user-defined courses/prereqs/coreqs}
\usage{
what_if(
  plan_of_study,
  courses_to_delete = NULL,
  reqs_to_delete = NULL,
  courses_to_add = NULL,
  prereqs_to_add = NULL,
  coreqs_to_add = NULL
)
}
\arguments{
\item{plan_of_study}{An igraph object created using the create_plan_of_study function}

\item{courses_to_delete}{An atomic vector listing the courses to delete (e.g., c("ENGR 101","ENGR 102"))}

\item{reqs_to_delete}{An atomic vector listing the pre/co requisties to delete (e.g.,
c("ENGR 101|ENGR 102","ENGL 101|ENGL 102")}

\item{courses_to_add}{An atomic vector listing the courses to add (e.g., c("ENGR 101","ENGR 102"))}

\item{prereqs_to_add}{An atomic vector listing the prerequisties to add
(e.g., c("ENGR 101","ENGR 102","ENGL 101","ENGL 102") where ENGR 101 is a prereq to ENGR 102)}

\item{coreqs_to_add}{An atomic vector listing the corequisties to add
(e.g., c("ENGR 101","ENGR 102","ENGL 101","ENGL 102") where ENGR 101 is a coreq with ENGR 102)}
}
\value{
A list containing the change in structural complexity, the new hypothetical structural complexity,
the list of hypothetical crucialities, and the new plan of study igraph object.
}
\description{
This function takes in a plan of study, then adjusts the plan of study subject to the following user-defined
hypotheticals.
}
